Run clippy on Windows too, and clear the lints that were hiding there
clippy ran only on ubuntu, so every `#[cfg(target_os = "windows")]` and
`#[cfg(not(unix))]` path was linted by nobody -- which is most of the
Windows-support code this fork exists for. `check` already matrixes both
platforms for exactly this reason ("Windows is a first-class target here and
the one that catches unix-only code"); clippy never got the same treatment.

The debt behind the gate turned out to be small and entirely cosmetic. Nothing
here changes behaviour:

- axon-config/src/shell.rs: two collapsible `if let` -> let-chains
- axon-file-utils/src/workspace_classifier.rs: one collapsible `if let`, and
  `map_or(false, ..)` -> `is_some_and(..)`
- axon-pager/src/app/mod.rs: collapsible `if let` in the stdin-mode restore
- axon-workspace/src/foreign_sessions/capability/mod.rs: needless `return`
- axon-crash-handler/src/handler.rs: `exception_code as i32` was an i32->i32
  no-op, since `EXCEPTION_RECORD.ExceptionCode` is already an NTSTATUS. Kept a
  comment so the removal does not read as an oversight -- si_code is passed to
  the same writer the unix arm feeds from libc.

Worth recording how the inventory was taken, because the obvious way is wrong:
`cargo clippy` replays cached diagnostics per unit, so a warm run silently
under-reports. A warm survey here returned 6 findings and omitted the
crash-handler cast entirely; another omitted the axon-pager one. Only a run in
a fresh CARGO_TARGET_DIR enumerates the real set. CI is unaffected -- it always
builds from scratch -- but do not trust a local warm clippy when deciding
whether a gate is safe to enable.

Two things this surfaced that are NOT addressed here: the Windows branch of
`is_platform_system_dir` has no test coverage (every test in that file is
`#[cfg(not(target_os = "windows"))]`), and `cargo test` cannot run on a Windows
host without NASM, because aws-lc-sys' build script requires it.

+++ crates/codegen/axon-crash-handler/src/handler.rs
@@ -620,7 +620,9 @@ mod win {
             }
 
             let signal = exception_to_signal(exception_code);
-            let si_code = exception_code as i32;
+            // `ExceptionCode` is already an NTSTATUS (i32), matching the
+            // libc `si_code` the Unix arm passes to the same writer.
+            let si_code = exception_code;
 
             // ExceptionInformation[1] holds the faulting address for ACCESS_VIOLATION.
             let si_addr = if exception_code == EXCEPTION_ACCESS_VIOLATION
